requirements.txt
langgraph==0.2.34
langchain-core==0.3.27
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
langchain-google-genai
